+++ test/e2e/specs/editor/plugins/inner-blocks-allowed-blocks.spec.js
@@ -114,7 +114,7 @@ test.describe( 'Allowed Blocks Setting on InnerBlocks', () => {
 		pageUtils,
 	} ) => {
 		await editor.canvas
-			.locator( 'role=button[name="Add default block"i]' )
+			.locator( 'role=document[name="Add default block"i]' )
 			.click();
 		await page.keyboard.type( '/Allowed Blocks Dynamic' );
 		await expect(